Move method to screen level
